@@ -1,6 +1,6 @@
 {
   "name": "payjp",
-  "version": "1.3.1",
+  "version": "1.3.2",
   "description": "PAY.JP node.js bindings",
   "main": "lib/index.js",
   "keywords": [
@@ -10,9 +10,9 @@
     "api"
   ],
   "scripts": {
-    "test": "./node_modules/.bin/mocha --compilers js:espower-babel/guess test/**/*.js",
-    "build": "./node_modules/babel-cli/bin/babel.js ./src --out-dir ./lib",
-    "lint": "./node_modules/eslint/bin/eslint.js src",
+    "test": "mocha --compilers js:espower-babel/guess",
+    "build": "babel ./src --out-dir ./lib",
+    "lint": "eslint src",
     "prepublish": "npm run build"
   },
   "repository": {
@@ -36,7 +36,7 @@
     "babel-plugin-add-module-exports": "^0.2.1",
     "babel-plugin-transform-es2015-modules-commonjs": "^6.4.5",
     "babel-preset-es2015": "^6.3.13",
-    "eslint": "^3.19.0",
+    "eslint": "^6.0.1",
     "espower-babel": "^4.0.1",
     "mocha": "^3.3.0",
     "power-assert": "^1.1.0",
